+++ Domain/Players/PlayerTablet.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
 using Domain.Exceptions;
+using Domain.Weapons;
 
 namespace Domain.Players
 {
