+++ Domain/PlayingCards/CarabineCard.cs
@@ -1,4 +1,4 @@
-﻿using Domain.Players;
+﻿using Domain.PlayingCards.Visitors;
 
 namespace Domain.PlayingCards
 {
@@ -8,7 +8,7 @@ namespace Domain.PlayingCards
         protected override bool EqualsCore(PlayingCard other) => other is CarabineCard;
 
         protected override int GetHashCodeCore() => typeof(CarabineCard).GetHashCode();
-        public override T Accept<T>(ICardVisitor<T> visitor) => visitor.Visit(this);
+        public override T Accept<T>(ILongTermCardVisitor<T> visitor) => visitor.Visit(this);
         
     }
 }
