+++ Domain/PlayingCards/VolcanicCard.cs
@@ -1,4 +1,4 @@
-﻿using Domain.Players;
+﻿using Domain.PlayingCards.Visitors;
 
 namespace Domain.PlayingCards
 {
@@ -7,8 +7,8 @@ namespace Domain.PlayingCards
         public override string Description => CardName.Volcanic;
         protected override bool EqualsCore(PlayingCard other) => other is VolcanicCard;
 
-        protected override int GetHashCodeCore() => typeof(Volcanic).GetHashCode();
-        public override T Accept<T>(ICardVisitor<T> visitor) => visitor.Visit(this);
+        protected override int GetHashCodeCore() => typeof(VolcanicCard).GetHashCode();
+        public override T Accept<T>(ILongTermCardVisitor<T> visitor) => visitor.Visit(this);
         
     }
 }
