+++ Domain/PlayingCards/WinchesterCard.cs
@@ -1,4 +1,4 @@
-﻿using Domain.Players;
+﻿using Domain.PlayingCards.Visitors;
 
 namespace Domain.PlayingCards
 {
@@ -6,8 +6,8 @@ namespace Domain.PlayingCards
     {
         public override string Description => CardName.Winchester;
         protected override bool EqualsCore(PlayingCard other) => other is WinchesterCard;
-        protected override int GetHashCodeCore() => typeof(Winchester).GetHashCode();
-        public override T Accept<T>(ICardVisitor<T> visitor) => visitor.Visit(this);
+        protected override int GetHashCodeCore() => typeof(WinchesterCard).GetHashCode();
+        public override T Accept<T>(ILongTermCardVisitor<T> visitor) => visitor.Visit(this);
         
     }
 }
