+++ Domain/Weapons/Carabine.cs
@@ -1,0 +1,13 @@
+namespace Domain.Weapons
+{
+    public class Carabine : Weapon
+    {
+        protected override bool EqualsCore(Weapon other)
+        {
+            return other is Carabine;
+        }
+
+        public override int Distance => 4;
+        public override bool MultipleBang => false;
+    }
+}
