+++ Domain/Weapons/Remington.cs
@@ -1,0 +1,13 @@
+namespace Domain.Weapons
+{
+    public class Remington : Weapon
+    {
+        protected override bool EqualsCore(Weapon other)
+        {
+            return other is Remington;
+        }
+
+        public override int Distance => 3;
+        public override bool MultipleBang => false;
+    }
+}
