+++ Domain/Weapons/Schofield.cs
@@ -1,0 +1,13 @@
+namespace Domain.Weapons
+{
+    public class Schofield : Weapon
+    {
+        protected override bool EqualsCore(Weapon other)
+        {
+            return other is Schofield;
+        }
+
+        public override int Distance => 2;
+        public override bool MultipleBang => false;
+    }
+}
