+++ Domain/Weapons/Winchester.cs
@@ -1,0 +1,13 @@
+namespace Domain.Weapons
+{
+    public class Winchester : Weapon
+    {
+        protected override bool EqualsCore(Weapon other)
+        {
+            return other is Winchester;
+        }
+
+        public override int Distance => 5;
+        public override bool MultipleBang => false;
+    }
+}
